Refactor export dialogs
