+++ frontend/src/components/ChatList.js
@@ -21,17 +21,8 @@ import {
   InputAdornment,
   CardActions,
   Tooltip,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions,
   FormControlLabel,
-  Checkbox,
-  DialogContentText,
   Switch,
-  Radio,
-  RadioGroup,
-  FormControl,
 } from '@mui/material';
 import FolderIcon from '@mui/icons-material/Folder';
 import CalendarTodayIcon from '@mui/icons-material/CalendarToday';
@@ -43,8 +34,9 @@ import ExpandLessIcon from '@mui/icons-material/ExpandLess';
 import SearchIcon from '@mui/icons-material/Search';
 import ClearIcon from '@mui/icons-material/Clear';
 import FileDownloadIcon from '@mui/icons-material/FileDownload';
-import WarningIcon from '@mui/icons-material/Warning';
 import { colors } from '../App';
+import { exportChat, exportPreferences } from '../utils/exportUtils';
+import { FormatSelectionDialog, ExportWarningDialog } from './ExportDialogs';
 
 const ChatList = () => {
   const [chats, setChats] = useState([]);
@@ -127,13 +119,7 @@ const ChatList = () => {
     fetchChats();
     
     // Check if user has previously chosen to not show the export warning
-    const warningPreference = document.cookie
-      .split('; ')
-      .find(row => row.startsWith('dontShowExportWarning='));
-    
-    if (warningPreference) {
-      setDontShowExportWarning(warningPreference.split('=')[1] === 'true');
-    }
+    setDontShowExportWarning(exportPreferences.getDontShowWarning());
   }, []);
 
   // Watch for changes to showDemoChats and refetch when it changes
@@ -216,10 +202,6 @@ const ChatList = () => {
   };
 
   // Handle format dialog selection
-  const handleFormatDialogOpen = () => {
-    setFormatDialogOpen(true);
-  };
-
   const handleFormatDialogClose = (confirmed) => {
     setFormatDialogOpen(false);
     
@@ -242,9 +224,7 @@ const ChatList = () => {
     
     // Save preference in cookies if "Don't show again" is checked
     if (dontShowExportWarning) {
-      const expiryDate = new Date();
-      expiryDate.setFullYear(expiryDate.getFullYear() + 1); // Cookie lasts 1 year
-      document.cookie = `dontShowExportWarning=true; expires=${expiryDate.toUTCString()}; path=/`;
+      exportPreferences.saveDontShowWarning(true);
     }
     
     // If confirmed, proceed with export
@@ -268,80 +248,14 @@ const ChatList = () => {
     setCurrentExportSession(sessionId);
     
     // Open format selection dialog first
-    handleFormatDialogOpen();
+    setFormatDialogOpen(true);
   };
 
   // Function to actually perform the export
   const proceedWithExport = async (sessionId, format) => {
-    try {
-      console.log(`Starting ${format.toUpperCase()} export for session:`, sessionId);
-      console.log(`Making API request to: /api/chat/${sessionId}/export?format=${format}`);
-      
-      const response = await axios.get(
-        `/api/chat/${sessionId}/export?format=${format}`,
-        { responseType: 'blob' }
-      );
-
-      const blob = response.data;
-      console.log('Received blob size:', blob ? blob.size : 0);
-
-      if (!blob || blob.size === 0) {
-        throw new Error('Received empty or invalid content from server');
-      }
-
-      // Ensure the blob has the correct MIME type
-      const mimeType = format === 'json' ? 'application/json;charset=utf-8' : 'text/html;charset=utf-8';
-      const typedBlob = blob.type ? blob : new Blob([blob], { type: mimeType });
-      console.log('Prepared typed blob, size:', typedBlob.size);
-
-      // --- Download Logic Start ---
-      const extension = format === 'json' ? 'json' : 'html';
-      const filename = `cursor-chat-${sessionId.slice(0, 8)}.${extension}`;
-      const link = document.createElement('a');
-      
-      // Create an object URL for the (possibly re-typed) blob
-      const url = URL.createObjectURL(typedBlob);
-      link.href = url;
-      link.download = filename;
-      
-      // Append link to the body (required for Firefox)
-      document.body.appendChild(link);
-      
-      // Programmatically click the link to trigger the download
-      link.click();
-      
-      // Clean up: remove the link and revoke the object URL
-      document.body.removeChild(link);
-      URL.revokeObjectURL(url);
-      console.log("Download initiated and cleanup complete");
-      // --- Download Logic End ---
-      
-    } catch (error) {
-      // ADDED: More detailed error logging
-      console.error('Detailed export error:', error);
-      if (error.response) {
-        // The request was made and the server responded with a status code
-        // that falls out of the range of 2xx
-        console.error('Error Response Data:', error.response.data);
-        console.error('Error Response Status:', error.response.status);
-        console.error('Error Response Headers:', error.response.headers);
-      } else if (error.request) {
-        // The request was made but no response was received
-        // `error.request` is an instance of XMLHttpRequest in the browser
-        console.error('Error Request:', error.request);
-      } else {
-        // Something happened in setting up the request that triggered an Error
-        console.error('Error Message:', error.message);
-      }
-      console.error('Error Config:', error.config);
-      
-      const errorMessage = error.response ? 
-        `Server error: ${error.response.status}` : 
-        error.request ? 
-        'No response received from server' : 
-        error.message || 'Unknown error setting up request';
-      alert(`Failed to export chat: ${errorMessage}`);
-    }
+    await exportChat(sessionId, format);
+    // Reset session after export completes
+    setCurrentExportSession(null);
   };
 
   if (loading) {
@@ -401,80 +315,21 @@ const ChatList = () => {
         </Box>
       </Box>
       
-      {/* Format Selection Dialog */}
-      <Dialog
+      {/* Use the reusable dialog components */}
+      <FormatSelectionDialog 
         open={formatDialogOpen}
-        onClose={() => handleFormatDialogClose(false)}
-        aria-labelledby="format-selection-dialog-title"
-      >
-        <DialogTitle id="format-selection-dialog-title" sx={{ display: 'flex', alignItems: 'center' }}>
-          <FileDownloadIcon sx={{ color: colors.highlightColor, mr: 1 }} />
-          Export Format
-        </DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            Please select the export format for your chat:
-          </DialogContentText>
-          <FormControl component="fieldset" sx={{ mt: 2 }}>
-            <RadioGroup
-              aria-label="export-format"
-              name="export-format"
-              value={exportFormat}
-              onChange={(e) => setExportFormat(e.target.value)}
-            >
-              <FormControlLabel value="html" control={<Radio />} label="HTML" />
-              <FormControlLabel value="json" control={<Radio />} label="JSON" />
-            </RadioGroup>
-          </FormControl>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={() => handleFormatDialogClose(false)} color="highlight">
-            Cancel
-          </Button>
-          <Button onClick={() => handleFormatDialogClose(true)} color="highlight" variant="contained">
-            Continue
-          </Button>
-        </DialogActions>
-      </Dialog>
+        onClose={handleFormatDialogClose}
+        exportFormat={exportFormat}
+        setExportFormat={setExportFormat}
+        colors={colors}
+      />
       
-      {/* Export Warning Modal */}
-      <Dialog
+      <ExportWarningDialog
         open={exportModalOpen}
-        onClose={() => handleExportWarningClose(false)}
-        aria-labelledby="export-warning-dialog-title"
-      >
-        <DialogTitle id="export-warning-dialog-title" sx={{ display: 'flex', alignItems: 'center' }}>
-          <WarningIcon sx={{ color: 'warning.main', mr: 1 }} />
-          Export Warning
-        </DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            Please make sure your exported chat doesn't include sensitive data such as API keys and customer information.
-          </DialogContentText>
-          <FormControlLabel
-            control={
-              <Checkbox
-                checked={dontShowExportWarning}
-                onChange={(e) => setDontShowExportWarning(e.target.checked)}
-              />
-            }
-            label="Don't show this warning again"
-            sx={{ mt: 2 }}
-          />
-        </DialogContent>
-        <DialogActions>
-          <Button 
-            onClick={() => handleExportWarningClose(false)} 
-            color="primary"
-            sx={{ color: 'white' }}
-          >
-            Cancel
-          </Button>
-          <Button onClick={() => handleExportWarningClose(true)} color="highlight" variant="contained">
-            Continue Export
-          </Button>
-        </DialogActions>
-      </Dialog>
+        onClose={handleExportWarningClose}
+        dontShowWarning={dontShowExportWarning}
+        setDontShowWarning={setDontShowExportWarning}
+      />
       
       {/* Search Bar */}
       <TextField
